third_party/blink/web_tests/http/tests/security/srcdoc-in-sandbox-cannot-access-parent-expected.txt

CONSOLE ERROR: Uncaught SecurityError: Sandbox access violation: Blocked a frame at "null" from accessing a frame at "http://127.0.0.1:8000".  The frame requesting access is sandboxed and lacks the "allow-same-origin" flag.
PASS
